Updated to support (EE-only) GCC v5.3.0. POC-only, not every problem is fixed.
Fixes to support (EE) GCC v5.3.0, in general:
1. Changed AS from ee-as to ee-gcc -c.
2. -mno-crt0 was replaced with just -nostartfiles.
3. As long ints are now 32-bit, all code that used long ints as a 64-bit data type were updated to use s64/u64.
4. Some assembler register names were changed, due to a different ABI being used. as_reg_compat.h was included in some files to help with compatibility, while others had to be edited.
5. GCC is now stricter in compliance with the C-standard.
6. There is still no Newlib port for this version of GCC, so C++ objects won't be built.
7. stdio.h was edited to use tamtypes.h instead of sys/types.h. ssize_t was replaced with size_t, and off_t is defined to be long int.
8. FIXME: when GCC has the calloc inlining optimization bug fixed, undo the workaround in alloc.c.

Some fixes were taken from Mega Man's fork of the ps2sdk (for the same purpose),
and from uyjulian's recent pull request.

+++ iop/tcpip/tcpips/src/ps2ips.c
@@ -440,17 +440,16 @@ static void do_select( void * rpcBuffer, int size )
 	}
 }
 
-// cmd should be 64 bits wide; I think.
 static void do_ioctlsocket( void *rpcBuffer, int size )
 {
 	int *ptr = rpcBuffer, ret;
 	int s;
-	unsigned int cmd;
+	long int cmd;
 	unsigned int argpv;
 	void *argp;
 
-	s = ((int*)_rpc_buffer)[0];
-	cmd = ((unsigned int*)_rpc_buffer)[1];
+	s = (int)((s32*)_rpc_buffer)[0];
+	cmd = (long int)((s32*)_rpc_buffer)[1];
 	argp = (void*)(((int*)_rpc_buffer)[2]);
 	if( argp )
 	{
